separate scene extensions
